requirements.txt
pycaret-ts-alpha==3.0.0.dev1638407153
scikit-learn==0.24.2
Pillow==8.4.0
pandas==1.3.4
autokeras==1.0.16
scikit-plot==0.3.7
seaborn==0.11.0
matplotlib==3.2.0
xgboost==1.3.3
lightgbm==3.2.1
shap==0.39.0
dill==0.3.2
pandas-profiling==2.12.0
explainerdashboard==0.3.6.2
streamlit-pandas-profiling==0.1.3
dask==2021.4.0
streamlit
openpyxl==3.0.7
category-encoders==2.2.2
statsmodels==0.12.1
numpy==1.19.5
dash-bootstrap-components==0.13.1